requirements.txt
anthropic
python-dotenv
ebooklib
beautifulsoup4
lxml
streamlit
